extract files of service file
