+++ wire-ios-data-model/Source/Utilis/EAR/EARService.swift
@@ -78,23 +78,6 @@ public protocol EARServiceInterface: AnyObject {
     func setInitialEARFlagValue(_ enabled: Bool)
 }
 
-public protocol EARServiceDelegate: AnyObject {
-
-    /// Prepare for the migration of existing database content.
-    ///
-    /// When the migration can be started, invoke the `onReady` closure.
-
-    func prepareForMigration(onReady: @escaping (NSManagedObjectContext) throws -> Void) rethrows
-
-}
-
-public enum EARServiceFailure: Error {
-
-    case cannotPerformMigration
-    case databaseKeyMissing
-
-}
-
 public class EARService: EARServiceInterface {
 
     // MARK: - Properties
@@ -504,36 +487,6 @@ public class EARService: EARServiceInterface {
                 block(context)
             }
         }
-    }
-
-}
-
-public struct EARPublicKeys {
-
-    public let primary: SecKey
-    public let secondary: SecKey
-
-    public init(
-        primary: SecKey,
-        secondary: SecKey
-    ) {
-        self.primary = primary
-        self.secondary = secondary
-    }
-
-}
-
-public struct EARPrivateKeys {
-
-    public let primary: SecKey?
-    public let secondary: SecKey
-
-    public init(
-        primary: SecKey?,
-        secondary: SecKey
-    ) {
-        self.primary = primary
-        self.secondary = secondary
     }
 
 }
